requirements.txt
boto3==1.26.142
requests==2.31.0
python-dotenv==1.0.0
lightgbm==3.3.3
optuna==3.1.1
numpy==1.24.3
scikit-learn==1.2.2